Try to resolve $BASH_SOURCE using readlink -e
If you use a symlink to git-hub command, it will fail to locate the bash+
dependency because it gets the symlink's directory instead of the
file's.
This attempt to resolve the symlink before doing so but since it uses
a GNU readlink options, it falls back to the previous behavior if it
fails.