Improve the pull-request command
This commit makes pull-request DWIM to find a target repository and
branch name.

Because specifying owner/repo or branch on the command line is ambiguous
as to whether it is source or target, we remove those options and rely
on the existing --remote and --branch options, plus new options --parent
and --base.