Fix for BSD/GNU `install` flag incompatibility.
In git-hub's makefile, `install` directives that use both -C (copy, do not change modification time if destination is the same) and -d (create directories as necessary) flags at the same time will fail on BSD and BSD derivatives. This is because BSD's `install` does not support simultaneous copy-without-modify and parent directory creation, for whatever reason. GNU's version of `install` does not have this problem, presumably because it doesn't need to be backwards-compatible with whichever original `install` prevented this.

This commit fixes the issue by removing the -C flag outright from the conflicting directives, as it does not add anything significant to the installation process.