Quote whole example manually-set (MAN)PATH values
In the Swim documentation document, `doc/git-hub.swim`, an example of
setting one’s `PATH` and `MANPATH` environment variables, so as to
include git-hub’s `lib` and `man` directories, manually (as opposed to
via the `init` script) showed only parts of the example values quoted,
which parts did not include the added elements, i.e., the paths to
git-hub’s `lib` and `man` directories.

These added paths may contain characters requiring quoting, so this
commit changes this example to instead have the whole values be quoted.