`init`: Fix `$GIT_HUB_ROOT` setup
In the `init` script, which configures a user’s shell environment for
easier access to `git-hub`, there is a section that sets the environment
variable `GIT_HUB_ROOT`, deriving its value, which supposedly equals the
file-path of the `git-hub` working tree, from the file-path of the
`init` script.

When the `init` script was moved out of `lib/git-hub.d/` into repository
top-level (i.e., up two levels in the filesystem hierarchy), the section
of it that derives the value of `GIT_HUB_ROOT` was not updated to
reflect the move, resulting in the file-path set as the value of
`GIT_HUB_ROOT` being equivalent to `<correct value>/../..`.

This commit fixes this issue.