.rc: Don't risk error on unset MANPATH
In the `.rc` initialization script, when mutating the `MANPATH` variable
such as to make our man-page discoverable via it, rather than
unconditionally attempting to substitute the old value of `MANPATH` into
its new value, instead substitute it only if it is set.

The purpose of this change is to avoid raising an error if --

  - attempting plain (unconditional) parameter expansion of an unset
    parameter is considered an error by the shell, as it may be due to
    shell options that the user may have set, or perhaps due to what
    shell the user may be using; and

  - `MANPATH` is unset, as it may well be (it's unset by default on my
    desktop and laptop computers (both running NixOS), an Arch Linux VPS
    I use, and, I'm told, on Ubuntu and FreeBSD).

I note that leaving a colon at the end of the value of `MANPATH` appears
to be necessary for man-pages installed at the OS level to still be
discoverable if `MANPATH` is normally unset.